✨  Add a new Skript Info command
This new skript info command show the latest version, and the latest
stable version if the latest is a prerelease. It gets its informations
from the cache, which is set every 10 minutes by the "skriptReleases"
task. We now also use the official Github API wrapper around their REST
API: Octokit. This allows us to have stronger typings for the responses.